Add files via upload
In this project on SQL, we analyzed the Parch and Posey dataset, which contains the sales data of different types of paper (gloss, standard) for a hypothetical paper company. The dataset is composed of various tables linked with a database schema.

I have provided a folder containing a SQL document and a Word document with all the questions and SQL code for answers. The project covers all advanced topics of SQL, and it consists of approximately 900 lines of code in total.

I have provided Dataset -  CSV files in the folder